requirements.txt
# absl-py==0.15.0
absl-py
# matplotlib==3.5.2
matplotlib
# ninja==1.10.2.3
ninja
# numpy==1.19.5
numpy
# pandas==1.3.3
pandas
# Pillow==9.2.0
Pillow
# ml-collections==0.1.0
ml-collections
# tensorboard==2.2.0
tensorboard
tensorflow_io
# tensorflow==2.2.0
tensorflow
# tensorflow-addons==0.12.0
tensorflow-addons
# tensorflow-datasets==4.5.2
tensorflow-datasets
# torch==1.10.2
torch
# torchmetrics==0.9.3
torchmetrics
# torchsummary==1.5.1
torchsummary
# torchvision==0.11.3
torchvision
# tqdm==4.62.3
tqdm
# typeguard==2.13.3
typeguard
# typing-extensions==3.7.4
typing-extensions

# added new module by yaguchi
jax
jaxlib
imageio
seaborn
wandb

pdf2image 
opencv-python-headless

pot